Use bindAttr when binding HTML attributes
Because we need to be able to update attributes
at a later date, attributes in Ember must be
emitted using the {{bindAttr}} helper (instead of
vanilla Handlebars, where we're just building a
string of HTML, and you can emit whatever you
want).

This fixes the bug where images were not
displaying.

For more info, see @wagenet's blog post:
http://www.emberist.com/2012/04/06/bind-and-bindattr.html